Cleanup Code and implement todo's
Code Cleanup using CppCheck and Microsoft Code Analysis, Now Iterating
Threads until all others are suspended, and patching IP if necessary.

+++ disasm-lib/misc.c
@@ -5,15 +5,15 @@ BOOL IsHexChar(BYTE ch)
 {
 	switch (ch)
 	{
-		case '0': case '1': case '2': case '3': 
-		case '4': case '5': case '6': case '7': 
-		case '8': case '9': 
-		case 'A': case 'a': case 'B': case 'b':
-		case 'C': case 'c': case 'D': case 'd':
-		case 'E': case 'e': case 'F': case 'f':
-			return TRUE;
-		default:
-			return FALSE;
+	case '0': case '1': case '2': case '3':
+	case '4': case '5': case '6': case '7':
+	case '8': case '9':
+	case 'A': case 'a': case 'B': case 'b':
+	case 'C': case 'c': case 'D': case 'd':
+	case 'E': case 'e': case 'F': case 'f':
+		return TRUE;
+	default:
+		return FALSE;
 	}
 }
 
@@ -27,7 +27,7 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 	if (!InputLength || !OutputLength) return NULL;
 	else *OutputLength = 0;
 
-	while (*Input && isspace(*Input)) { Input++; InputLength--; }
+	while (*Input && isspace((unsigned)*Input)) { Input++; InputLength--; }
 	if (!*Input) return NULL;
 	if (Input[0] == '\"') { Input++; InputLength--; }
 	p = (BYTE *)strchr(Input, '\"');
@@ -37,7 +37,7 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 	{
 		for (i = 0; i < InputLength; i += 3)
 		{
-			while (i < InputLength && isspace(Input[i])) i++; // skip over extra space, \r, and \n
+			while (i < InputLength && isspace((unsigned)Input[i])) i++; // skip over extra space, \r, and \n
 			if (i >= InputLength) break;
 
 			if (!IsHexChar(Input[i]))
@@ -46,13 +46,13 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 				goto abort;
 			}
 
-			if (i+1 >= InputLength || !Input[i+1])
+			if (i + 1 >= InputLength || !Input[i + 1])
 			{
 				//fprintf(stderr, "ERROR: hex string terminates unexpectedly at offset %lu (0x%04x)\n", i+1, i+1);
 				goto abort;
 			}
 
-			if (i+2 < InputLength && Input[i+2] && !isspace(Input[i+2]))
+			if (i + 2 < InputLength && Input[i + 2] && !isspace((unsigned)Input[i + 2]))
 			{
 				//fprintf(stderr, "ERROR: Hex string is malformed at offset %lu (0x%04x)\n", i, i);
 				//fprintf(stderr, "Found '%c' (0x%02x) instead of space\n", Input[i+2], Input[i+2]);
@@ -68,19 +68,19 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 			goto abort;
 		}
 
-		ByteString = malloc(ByteCount+1);
+		ByteString = malloc(ByteCount + 1);
 		if (!ByteString)
 		{
 			//fprintf(stderr, "ERROR: failed to allocate %lu bytes\n", ByteCount);
 			goto abort;
 		}
-			
-		memset(ByteString, 0, ByteCount+1);
+
+		memset(ByteString, 0, ByteCount + 1);
 		for (i = 0, j = 0; j < ByteCount; i += 3, j++)
-		{			
-			while (isspace(Input[i])) i++; // skip over extra space, \r, and \n
+		{
+			while (isspace((unsigned)Input[i])) i++; // skip over extra space, \r, and \n
 			temp_byte[0] = Input[i];
-			temp_byte[1] = Input[i+1];
+			temp_byte[1] = Input[i + 1];
 			temp_byte[2] = 0;
 			ByteString[j] = (BYTE)strtoul(temp_byte, NULL, 16);
 		}
@@ -89,7 +89,7 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 	{
 		for (i = 0; i < InputLength; i += 2)
 		{
-			if (Input[i] != '\\' || (Input[i+1] != 'x' && Input[i+1] != '0'))
+			if (Input[i] != '\\' || (Input[i + 1] != 'x' && Input[i + 1] != '0'))
 			{
 				//fprintf(stderr, "ERROR: invalid hex character at offset %lu (0x%04x)\n", i, i);
 				goto abort;
@@ -101,7 +101,7 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 				//fprintf(stderr, "ERROR: invalid hex character at offset %lu (0x%04x)\n", i, i);
 				goto abort;
 			}
-			if (i+1 >= InputLength || !Input[i+1])
+			if (i + 1 >= InputLength || !Input[i + 1])
 			{
 				//fprintf(stderr, "ERROR: hex string terminates unexpectedly at offset %lu (0x%04x)\n", i+1, i+1);
 				goto abort;
@@ -116,19 +116,19 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 			goto abort;
 		}
 
-		ByteString = malloc(ByteCount+1);
+		ByteString = malloc(ByteCount + 1);
 		if (!ByteString)
 		{
 			//fprintf(stderr, "ERROR: failed to allocate %lu bytes\n", ByteCount);
 			goto abort;
 		}
-			
-		memset(ByteString, 0, ByteCount+1);
+
+		memset(ByteString, 0, ByteCount + 1);
 		for (i = j = 0; j < ByteCount; i += 2, j++)
 		{
 			i += 2;
 			temp_byte[0] = Input[i];
-			temp_byte[1] = Input[i+1];
+			temp_byte[1] = Input[i + 1];
 			temp_byte[2] = 0;
 			ByteString[j] = (BYTE)strtoul(temp_byte, NULL, 16);
 		}
@@ -137,12 +137,12 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 	{
 		for (i = 0; i < InputLength; i += 2)
 		{
-				if (!IsHexChar(Input[i]))
+			if (!IsHexChar(Input[i]))
 			{
 				//fprintf(stderr, "ERROR: invalid hex character at offset %lu (0x%04x)\n", i, i);
 				goto abort;
 			}
-			if (i+1 >= InputLength || !Input[i+1])
+			if (i + 1 >= InputLength || !Input[i + 1])
 			{
 				//fprintf(stderr, "ERROR: hex string terminates unexpectedly at offset %lu (0x%04x)\n", i+1, i+1);
 				goto abort;
@@ -157,18 +157,18 @@ BYTE *HexToBinary(char *Input, DWORD InputLength, DWORD *OutputLength)
 			goto abort;
 		}
 
-		ByteString = malloc(ByteCount+1);
+		ByteString = malloc(ByteCount + 1);
 		if (!ByteString)
 		{
 			//fprintf(stderr, "ERROR: failed to allocate %lu bytes\n", ByteCount);
 			goto abort;
 		}
-			
-		memset(ByteString, 0, ByteCount+1);
+
+		memset(ByteString, 0, ByteCount + 1);
 		for (i = 0, j = 0; j < ByteCount; i += 2, j++)
 		{
 			temp_byte[0] = Input[i];
-			temp_byte[1] = Input[i+1];
+			temp_byte[1] = Input[i + 1];
 			temp_byte[2] = 0;
 			ByteString[j] = (BYTE)strtoul(temp_byte, NULL, 16);
 		}
@@ -182,4 +182,3 @@ abort:
 	if (ByteString) free(ByteString);
 	return NULL;
 }
-
